flight_renamer/pyproject.toml
[project]
name = "fight-renamer"
version = "0.1.0"
description = "Rename output drone imagery into a structured directory, with metadata."
authors = [
    {name = "spwoodcock", email = "sam.woodcock@protonmail.com"},
]
dependencies = [
    "pillow>=10.3.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
